Move _expand_wfss_table to exp_to_source from flat_multispec module as Ned Molter requested
